use uint16 instead of float to spwm freq, slove the problem of float calculation error, and add freq step to 3

+++ program/Core/Inc/main.h
@@ -46,8 +46,8 @@ extern "C" {
 
 /* Exported macro ------------------------------------------------------------*/
 /* USER CODE BEGIN EM */
-extern float target_spwm_freq;
-extern float spwm_freq_step;
+extern uint16_t target_spwm_freq;
+extern uint8_t spwm_freq_step;
 /* USER CODE END EM */
 
 /* Exported functions prototypes ---------------------------------------------*/
